implemented the user service layer

+++ app/models/UserRole.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.Json
 
-case class UserRole(id: Int, name: String, level: Byte) {
+case class UserRole(id: Int, name: String, level: Int) {
 
 }
